requirements.txt
torch
torchvision
transformers
numpy
matplotlib
pillow
tqdm
torchmetrics
jupyter
diffusers
safetensors
accelerate
